test(host): match TODO's full strict warning set, default it on
1.2 follow-up - upgrade ENABLE_STRICT from a bare -Wall -Wextra -Werror
to the full set the implementation TODO specifies (-Wpedantic -Wshadow
-Wconversion -Wsign-conversion -Wformat=2 -Wundef -Wcast-qual
-Wstrict-prototypes) and default it ON so regressions surface
immediately; --no-strict opts out.

Fixes the handful of real issues this surfaced:
- esp_crt_bundle_attach's host mock returned void* instead of
  esp_err_t, and esp_http_client_config_t.crt_bundle_attach was typed
  void* instead of a function pointer — masked a pedantic function
  pointer/object pointer mismatch that doesn't reflect the real
  ESP-IDF API shape.
- pdMS_TO_TICKS() divided a possibly-signed ms by an unsigned
  constant without an explicit cast.
- a handful of int-to-size_t/int8_t narrowing casts in bt_link.c,
  radio.c, and two test files that were implicit before.

+++ esp_i2s_source/components/bt_link/bt_link.c
@@ -72,7 +72,7 @@ static void bt_link_task(void *arg)
                 bt_link_session_begin(&s_session, req->cmd);
                 char out[BT_LINK_LINE_MAX + 2];
                 int n = snprintf(out, sizeof(out), "%s\r\n", req->cmd);
-                uart_write_bytes(BT_LINK_UART, out, n);
+                uart_write_bytes(BT_LINK_UART, out, (size_t)n);
             }
         }
